# app/src/test/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: ""
  description: Generated server url
paths:
  /api/workspaces/{workspaceId}:
    put:
      tags:
      - workspaces-api
      operationId: editWorkspace
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditWorkspaceDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDto"
...
  /api/workspaces/{workspaceId}/incomes/{incomeId}:
    get:
      tags:
      - incomes-api
      operationId: getIncome
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: incomeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeDto"
    put:
      tags:
      - incomes-api
      operationId: updateIncome
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: incomeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditIncomeDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeDto"
  /api/workspaces/{workspaceId}/income-tax-payments/{taxPaymentId}:
    get:
      tags:
      - income-tax-payments-api
      operationId: getTaxPayment
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taxPaymentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeTaxPaymentDto"
    put:
      tags:
      - income-tax-payments-api
      operationId: updateTaxPayment
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taxPaymentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditIncomeTaxPaymentDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeTaxPaymentDto"
...
  /api/users/{userId}:
    get:
      tags:
      - users-api
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUserDto"
    put:
      tags:
      - users-api
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUserDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersApiUpdateUserErrors"
...
  /api/workspaces:
    get:
      tags:
      - workspaces-api
      operationId: getWorkspaces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceDto"
    post:
      tags:
      - workspaces-api
      operationId: createWorkspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDto"
...
  /api/workspaces/{workspaceId}/income-tax-payments:
    get:
      tags:
      - income-tax-payments-api
      operationId: getTaxPayments
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiPageIncomeTaxPaymentDto"
    post:
      tags:
      - income-tax-payments-api
      operationId: createTaxPayment
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditIncomeTaxPaymentDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeTaxPaymentDto"
...
  /api/user-activation-tokens:
    post:
      tags:
      - user-activation-tokens-api
      operationId: createToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserActivationTokenRequestDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivationTokenDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivationTokensApiCreateTokenErrors"
  /api/user-activation-tokens/{token}/activate:
    post:
      tags:
      - user-activation-tokens-api
      operationId: activateUser
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserActivationRequestDto"
        required: true
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivationTokensApiActivateUserErrors"
  /api/shared-workspaces:
    get:
      tags:
      - workspaces-api
      operationId: getSharedWorkspaces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceDto"
    post:
      tags:
      - workspaces-api
      operationId: saveSharedWorkspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSharedWorkspaceRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDto"
...
  /api/user-activation-tokens/{token}:
    get:
      tags:
      - user-activation-tokens-api
      operationId: getToken
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivationTokenDto"
  /api/user-activation-tokens/users/{userId}:
    get:
      tags:
      - user-activation-tokens-api
      operationId: getTokenByUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivationTokenDto"
...
components:
  schemas:
...
    EditWorkspaceDto:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 0
          type: string
    WorkspaceDto:
      required:
      - defaultCurrency
      - editable
      - multiCurrencyEnabled
      - name
      - taxEnabled
      - version
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        name:
          type: string
        taxEnabled:
          type: boolean
        multiCurrencyEnabled:
          type: boolean
        defaultCurrency:
          type: string
        editable:
          type: boolean
...
    EditIncomeDto:
      required:
      - currency
      - dateReceived
      - originalAmount
      - title
      - useDifferentExchangeRateForIncomeTaxPurposes
      type: object
      properties:
        category:
          type: integer
          format: int64
        dateReceived:
          type: string
          format: date
        title:
          type: string
        currency:
          type: string
        originalAmount:
          type: integer
          format: int64
        convertedAmountInDefaultCurrency:
          type: integer
          format: int64
        useDifferentExchangeRateForIncomeTaxPurposes:
          type: boolean
        incomeTaxableAmountInDefaultCurrency:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        notes:
          maxLength: 1024
          minLength: 0
          type: string
        generalTax:
          type: integer
          format: int64
        linkedInvoice:
          type: integer
          format: int64
    IncomeAmountsDto:
      type: object
      properties:
        originalAmountInDefaultCurrency:
          type: integer
          format: int64
        adjustedAmountInDefaultCurrency:
          type: integer
          format: int64
    IncomeDto:
      required:
      - attachments
      - convertedAmounts
      - currency
      - dateReceived
      - id
      - incomeTaxableAmounts
      - originalAmount
      - status
      - timeRecorded
      - title
      - useDifferentExchangeRateForIncomeTaxPurposes
      - version
      type: object
      properties:
        category:
          type: integer
          format: int64
        title:
          type: string
        timeRecorded:
          type: string
          format: date-time
        dateReceived:
          type: string
          format: date
        currency:
          type: string
        originalAmount:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        notes:
          type: string
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - FINALIZED
          - PENDING_CONVERSION
          - PENDING_CONVERSION_FOR_TAXATION_PURPOSES
        linkedInvoice:
          type: integer
          format: int64
        generalTax:
          type: integer
          format: int64
        generalTaxRateInBps:
          type: integer
          format: int32
        generalTaxAmount:
          type: integer
          format: int64
        convertedAmounts:
          $ref: "#/components/schemas/IncomeAmountsDto"
        incomeTaxableAmounts:
          $ref: "#/components/schemas/IncomeAmountsDto"
        useDifferentExchangeRateForIncomeTaxPurposes:
          type: boolean
    EditIncomeTaxPaymentDto:
      required:
      - amount
      - datePaid
      - title
      type: object
      properties:
        datePaid:
          type: string
          format: date
        reportingDate:
          type: string
          format: date
        amount:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        notes:
          type: string
        title:
          type: string
    IncomeTaxPaymentDto:
      required:
      - amount
      - attachments
      - datePaid
      - id
      - reportingDate
      - timeRecorded
      - title
      - version
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        title:
          type: string
        timeRecorded:
          type: string
          format: date-time
        datePaid:
          type: string
          format: date
        reportingDate:
          type: string
          format: date
        amount:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        notes:
          type: string
...
    UpdateUserRequestDto:
      required:
      - userName
      type: object
      properties:
        userName:
          maxLength: 255
          minLength: 0
          type: string
    PlatformUserDto:
      required:
      - activated
      - admin
      - userName
      - version
      type: object
      properties:
        userName:
          type: string
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        admin:
          type: boolean
        activated:
          type: boolean
...
    CreateWorkspaceDto:
      required:
      - defaultCurrency
      - name
      type: object
      properties:
        name:
          type: string
        defaultCurrency:
          type: string
...
    CreateUserActivationTokenRequestDto:
      required:
      - userId
      type: object
      properties:
        userId:
          type: integer
          format: int64
    UserActivationTokenDto:
      required:
      - expiresAt
      - token
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
    UserActivationRequestDto:
      required:
      - password
      type: object
      properties:
        password:
          maxLength: 100
          minLength: 0
          type: string
    SaveSharedWorkspaceRequestDto:
      required:
      - token
      type: object
      properties:
        token:
          type: string
...
    ApiPageIncomeTaxPaymentDto:
      required:
      - data
      - pageNumber
      - pageSize
      - totalElements
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        data:
          type: array
          items:
            $ref: "#/components/schemas/IncomeTaxPaymentDto"
...
    UserActivationTokensApiActivateUserErrors:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          enum:
          - TokenExpired
        message:
          type: string
    UserActivationTokensApiCreateTokenErrors:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          enum:
          - UserAlreadyActivated
        message:
          type: string
...
    UsersApiUpdateUserErrors:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          enum:
          - UserAlreadyExists
        message:
          type: string
...